+++ src/components/network-modification-table/renderers/select-header-cell.tsx
@@ -8,7 +8,7 @@
 import React, { useCallback } from 'react';
 import { Checkbox } from '@mui/material';
 import { Table } from '@tanstack/react-table';
-import { ComposedModificationMetadata } from '../utils';
+import { ComposedModificationMetadata } from '../../../utils';
 
 interface SelectHeaderCellProps {
     table: Table<ComposedModificationMetadata>;
