+++ src/utils/types/index.ts
@@ -21,3 +21,4 @@ export * from './loadflow.type';
 export * from './sensitivity-analysis.type';
 export * from './dynamic-margin-calculation.type';
 export * from './network-modification-types';
+export * from './network-modification-metadata';
